v2-new/api/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
tensorflow>=2.13.0
pystow>=0.5.0
jpype1>=1.4.0
numpy<2
pandas>=2.0.0
rdkit-pypi>=2022.9.5
nltk>=3.8.0
py2opsin>=1.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
psutil>=5.9.0
typing-extensions>=4.8.0
pydantic>=2.0.0 
ray[default]>=2.9.0 